Restore HTML for comments, embeddable convention
The include for HTML syntax was removed in 07c16c7d77. The `@scalaHtml`
cluster reference was put back in 3203d67af3 but the include wasn’t.

The `main_syntax` convention is followed by syntax definitions that
want to support being embedded, e.g. to be highlighted in fenced code
blocks in a Markdown file. html.vim is one example of the pattern.

All includes need to be done before any of our Scala syntax
configuration, and then their `b:current_syntax` unlet.