Remove help plugin stuff
IMO this just doesn't belong here, it ought to be a separate plugin or
perhaps in Derek's personal ~/.vim/after/plugin :-) It doesn't even
appear to apply to the current help file format for this project.

I also didn't understand the change in 65e924f9419 from #6 that put the
`JustifyCurrentLine` mapping in `ftplugin/scala.vim` -- the function is
for help files yet it was mapped buffer-local in Scala files. Goodbye.

Also while I was looking at mappings, put the current ones in Vim doc.