Add standard plugin guards, modelines
Copping out of ‘compatible’ mode by saying we don’t support it, because
who knows if we do?! Who runs Vim that way for real work? This might
have to be changed if there is ever an effort to get some of this
plugin included in Vim’s standard runtime files, but until then, this
spares us from worrying about crap like `:h use-cpo-save`.

Some files don’t actually currently comply with the modelines, but I
didn’t want to introduce a bunch of re-indenting in this commit, but
set standards going forward that match most of the current code.