Set correct highlight group for escaped characters
`scalaEscapedChar` should be set to `Special`, not `Function`.

To recreate the issue that this is fixing run:

```sh
echo "a\nb" > ~/foo.txt
```

Then use `:setfiletype` to change the filetype to various other
languages like c, java, perl, python, ruby, javascript, and finally
scala.  Notice that the `\n` is a different color only in scala.  It is
the color used for functions and it is the only outlier here.

This commit fixes that.